cleaned up our temporary components and removed the outdated ones, also updated the names of the components to better label their functionality
