+++ src/app/dashboard/components/AppointmentPanel.js
@@ -1,6 +1,6 @@
 "use client";
 
-export default function AppointmentDashDash() {
+export default function AppointmentPanel() {
   return (
     <div className="flex flex-col gap-2 p-5 md:p-0">Coming Soon...</div>
   );
