+++ src/app/dashboard/page.js
@@ -1,8 +1,7 @@
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "@/pages/api/auth/[...nextauth]";
 
-import UserPanel from "./components/tempUserPanel";
-import ProfilePanel from "./components/tempProfilePanel";
+import DashboardNavPanel from "./components/DashboardNavPanel";
 
 export default async function Dashboard() {
   const session = await getServerSession(authOptions);
@@ -11,7 +10,7 @@ export default async function Dashboard() {
     return (
       <main className="flex-col my-8 lg:my-16">
         <div className="mb-8 items-stretch mx-auto max-w-6xl gap-5 lg:flex">
-          <UserPanel {...session} />
+          <DashboardNavPanel {...session} />
         </div>
       </main>
     );
